Generate contents.json in the output folder
Instead of replacing the contents of the whole site in each page (which
could get really really big), generate a contents.json on the output
folder that can be fetched by the frontend JS to do interesting things
with.

That way we use the unused variable for the output file name, and we
cleaned up the code for the generation a little bit.